# [Discussion] remove mutex in batch span processor.