Instead of using mutexes, we just clone the sender.

The performance improves

```
time:   [1.5888 ms 1.5997 ms 1.6113 ms]
change: [-44.416% -43.990% -43.497%] (p = 0.00 < 0.05)
```

related #520