@jtescher 2021-04-24
@TommyCpp you mentioned in #520 that cloning would increase capacity, are span limits still being enforced with this change?