@TommyCpp 2021-04-24
> are span limits still being enforced with this change

I thought so. But upon closer examination, I found that the future bound channel guaranteed to allow every sender to send at least once. Thus, the `try_send` will always succeed for a newly created `Sender`. That means we will not get errors when the channel is full. 

Also based on the discussion here https://github.com/rust-lang/futures-rs/issues/403#issuecomment-284774096. The capacity of the bound channel is `buffer size +  number of senders there have ever been` which will lead to OOM in our use cases. Although I found this to be surprising.