@TommyCpp 2021-04-25
> new thread like the current simple processor / crossbeam sync channel that loops sending to the future mpsc channel or drops if that channel is full?

Can try that. Crossbeam channel's `try_send` gives us the desired behavior and it takes a `&self`. We could spawn a new thread to "relay" the message from the main task/thread to the export task/thread. The `max_queue_size` can be enforced in the crossbeam channel. I'm not sure if that's going to be a performance improvement, but I can bench it when I finish a POC.