@jtescher 2021-04-26
> unbound channel and maintain an internal state ourselves

What about an unbounded crossbeam channel, keeping the inner bounded channel for the limits?