@TommyCpp 2021-04-26
> > unbound channel and maintain an internal state ourselves
> 
> What about an unbounded crossbeam channel, keeping the inner bounded channel for the limits?

I don't think we need a dedicated thread or crossbeam channel in this case. Suppose we defined the channel like
```rust
struct Sender {
    unbounded_sender: UnbounedSender<Msg>,
    state: Arc<State>,
}

struct Receiver {
   unbounded_receiver: UnboundedReceiver<Msg>,
   state: Arc<State>
}

struct State {
   elements_num: AtomicNumber,
   ...
}
```

We can then define a `try_send` method with `&self` rather than `&mut self`. We can enforce the span limit with `State`.