@TommyCpp 2021-04-29
I found the `async_channel` crate seems to be a good choice when I research the implementation of the channel. Its sender has a `try_send` method that takes `&self`. I did a POC with it and most results look good except it performed worse if there wasn't any concurrency

```
BatchSpanProcessor/with 1 concurrent task                        
                        time:   [936.25 us 969.14 us 1.0046 ms]
                        change: [+9.6398% **+21.109%** +31.546%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 1 outliers among 50 measurements (2.00%)
  1 (2.00%) high severe

BatchSpanProcessor/with 2 concurrent task                        
                        time:   [744.50 us 760.84 us 777.97 us]
                        change: [-39.810% -33.943% -28.101%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 50 measurements (4.00%)
  2 (4.00%) high severe

BatchSpanProcessor/with 4 concurrent task                        
                        time:   [854.81 us 863.26 us 872.00 us]
                        change: [-46.646% -44.508% -41.412%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 50 measurements (4.00%)
  1 (2.00%) high mild
  1 (2.00%) high severe

BatchSpanProcessor/with 8 concurrent task                        
                        time:   [1.4787 ms 1.4834 ms 1.4889 ms]
                        change: [-50.795% -48.749% -46.637%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 50 measurements (22.00%)
  1 (2.00%) low severe
  4 (8.00%) low mild
  6 (12.00%) high severe

BatchSpanProcessor/with 16 concurrent task                        
                        time:   [2.5739 ms 2.5812 ms 2.5890 ms]
                        change: [-51.467% -49.728% -48.071%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 50 measurements (12.00%)
  2 (4.00%) low mild
  1 (2.00%) high mild
  3 (6.00%) high severe

BatchSpanProcessor/with 32 concurrent task                        
                        time:   [4.4602 ms 4.5027 ms 4.5489 ms]
                        change: [-49.916% -48.948% -47.841%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 50 measurements (8.00%)
  2 (4.00%) high mild
  2 (4.00%) high severe

```