@jtescher 2021-04-29
@TommyCpp how do the existing span start/end overhead benchmarks compare? E.g. if you change the group function to 

```rust
fn trace_benchmark_group<F: Fn(&sdktrace::Tracer)>(c: &mut Criterion, name: &str, f: F) {
    let rt = tokio::runtime::Runtime::new().unwrap();
    rt.block_on(async {
        let mut group = c.benchmark_group(name);

        group.bench_function("always-sample", |b| {
            let provider = sdktrace::TracerProvider::builder()
                .with_config(sdktrace::config().with_sampler(sdktrace::Sampler::AlwaysOn))
                .with_default_batch_exporter(VoidExporter, opentelemetry::runtime::Tokio)
                .build();
            let always_sample = provider.get_tracer("always-sample", None);

            b.iter(|| f(&always_sample));
        });

        group.bench_function("never-sample", |b| {
            let provider = sdktrace::TracerProvider::builder()
                .with_config(sdktrace::config().with_sampler(sdktrace::Sampler::AlwaysOff))
                .with_default_batch_exporter(VoidExporter, opentelemetry::runtime::Tokio)
                .build();
            let never_sample = provider.get_tracer("never-sample", None);
            b.iter(|| f(&never_sample));
        });

        group.finish();
    });
}
```

With the mutex, the batch processor seems to have about twice the overhead on the application process, wondering how that changes with the various approaches you've tried.