@TommyCpp 2021-05-18
Did another round of benchmarking

| Benchmark | Baseline(Current implementation) | mpmc channel | new thread| Tokio channel|
|------------|---------------|-----------|----------------| ---------|
|BSP/1 task| 1.2ms | 1.4ms ( +14%) | 664us (-47%) | 759us(-8%)|
|BSP/2 tasks| 1.3ms |  930 us (-31%) | 865us(-38%) | 664.78us(-47%) |
|BSP/4 tasks| 2.0ms | 1.07ms(-45%) | 1.16ms(-41%) | 865.06us(-38%) |
|BSP/8 tasks| 3.5ms|  1.69ms(-52%) | 1.7ms(-51%) | 1.30ms(-41%) |
|BSP/16 tasks| 6.3ms| 2.68ms(-57%) | 3.1ms(-50%) | 1.7181ms(-51%) |
|BSP/32 tasks| 10.6ms | 4.7ms(-55%) | 5.0ms(-49%)| 3.13ms(-50%) |
|start-end-span/always-sample| 941.14ns |429ns (-54%) | 436ns(-50%) | 454.97ns(-51%) |
|start-end-span/never-sample| 170.07ns |166.78ns(-3%)| 150ns(-11%) | 149.71ns(-11.987%) | 
|start-end-span-4-attrs/always-sample| 1.8034us | 1.08us(-41%) | 1.23us(-33%) | 1.18us(-36%) |
|start-end-span-4-attrs/never-sample| 187.49ns | 179.95ns(-4%) | 186.63ns(+0.89%) | 169.64ns(-9%) |
|start-end-span-8-attrs/always-sample| 2.38us | 1.7092us(-28%) | 1.9882us(-16%) | 1.7026us(-28%) | 
|start-end-span-8-attrs/never-sample| 214ns | 231.45ns(+5%) | 223.82ns(+3.9%) | 211.24ns(-1.9%) |
|start-end-span-all-attr-types/always-sample| 2.0105us | 1.63(-19%) | 1.79us(-11%) | 1.45us(-28.6%) |
|start-end-span-all-attr-types/never-sample| 195.05us | 200.47ns(+1.1%) | 218ns(+11%) | 186.63ns(-4%) | 
|start-end-span-all-attr-types-2x/always-sample| 3.8546us | 2.6451us(-30.9%) | 3.34us(-11%) | 2.81us(-27%) |
| start-end-span-all-attr-types-2x/never-sample| 259.07ns | 276.76(+3.92%) | 254.04ns(-2.7%) | 260.54ns(-7%) |


Using a new thread is probably a good start. But I am still worried about the possibility of OOM here. If we have multiple producers it could in theory OOM because the relay thread cannot handle the incoming volume right?